requirements.txt
flask==2.3.3
werkzeug==2.3.7
fastapi>=0.100.0
uvicorn==0.23.2
pymongo==4.5.0
python-dotenv==1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.0.0.post2
requests==2.31.0
typing_extensions>=4.7.1
gunicorn==21.2.0
